Move MetaResult from base.meta to base.base.
